chore(wizard-refactor): extract tools/create_course (the elephant, ~1360 lines)
Last and largest tool extract. Pinned by 2 dispatcher tests (graph guard
HARD-FAIL + #607 unlink invariant on reuse path). 61/61 wizard tests green.

Monolith is now 86 lines: thin dispatcher + WizardToolResult re-export.
create_course.ts stays as a single file per the handoff — internal
re-split (graph guard / domain resolution / reuse vs new path / behavior
targets / enrollment / lesson plan) is a follow-on PR.

One relative import promoted to canonical @/ alias:
  await import("./wizard-ai-output-guard")
    → await import("@/lib/chat/wizard-ai-output-guard")
since the file moved deeper into the new directory.
